refactor with util module
